requirements.txt
streamlit==1.32.2
pandas==2.2.0
numpy==1.26.4
openpyxl==3.1.2
pytest==8.0.0
plotly==5.19.0
psycopg2-binary==2.9.9
